refactor(#7): restructure to match new outline
Fixes #7
