+++ .gitignore
@@ -18,4 +18,4 @@ cache
 build
 card_cache
 *.egg-info
-.testmondata
+.testmondata*
