+++ memorymarker/persist_questions/markdown.py
@@ -1,6 +1,6 @@
 from pathlib import Path
 
-from memorymarker.domain.highlights_to_questions import QAPrompt
+from memorymarker.question_generator.question_generator import QAPrompt
 
 SAVE_DIR = Path("card_cache")
